refactoring tests cases common functions to test_case_base
