+++ fullhouse/test/test_add_house_members.py
@@ -1,10 +1,8 @@
-from django.test import TestCase
 from django.test.client import Client
-from django.contrib.auth.models import User
 from django.core import mail
-import re
+import test_case_base
 
-class TestAddHouseMembers(TestCase):
+class TestAddHouseMembers(test_case_base.TestCaseBase):
     def setUp(self):
         self.client     = Client()
         self.email      = 'alice@eatallthecake.com'
@@ -14,67 +12,7 @@ class TestAddHouseMembers(TestCase):
         
         self.createUser(self.email, self.password)
         self.createUser(self.email2, self.password2)
-        self.loginUser(self.email, self.password);
-      
-    def createUser(self, email, password):
-        '''
-        Utility function creates a new user with given  email and pwd
-        '''
-        # clear email client 
-        mail.outbox = []
-        
-        # Create registration data
-        user_registration_data = {
-            'email': email,
-            'password1': password,
-            'password2': password,
-        }
-      
-        # register user
-        with self.settings(
-            EMAIL_BACKEND='django.core.mail.backends.locmem.EmailBackend'
-            ):
-            # registration redirects to dashboard on success
-            response = self.client.post('/accounts/register/', follow=True,
-                data=user_registration_data)
-            self.assertEqual( response.redirect_chain,
-            [('http://testserver/accounts/register/complete/', 302)]
-        )
-        self.assertEqual(response.status_code, 200)
-
-
-        # now get the activation email and trigger activation for the 
-        # user
-        email = mail.outbox[0]
-        match = re.search(r'/accounts/activate/([a-f0-9]{40})/',
-                          str(email.message()))
-        if not match:
-            raise Exception('failed to find key in activation email')
-        activation_key = match.group(1)
-
-        response = self.client.get('/accounts/activate/%s/' % activation_key,
-                                   follow=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            response.redirect_chain,
-            [('http://testserver/accounts/activate/complete/', 302)]
-        )
-        
-    def loginUser(self, eml, pwd):
-        response = self.client.post('/accounts/login/',
-            data={'email': eml, 'password': pwd},
-            follow=True
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            response.redirect_chain,
-            [('http://testserver/dashboard/', 302)]
-        )
-        ## find user based on email 
-        user = User.objects.get(email = eml)
-        self.assertEqual(self.client.session['_auth_user_id'], user.pk)
-        
-        
+        self.loginUser(self.email, self.password);  
 
     def test_create_house(self):
